circuits/requirements.txt
aiocache==0.12.2
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anvil-web3==0.0.4
anyio==4.4.0
astroid==3.2.2
async-cache==1.1.1
attrs==23.2.0
autopep8==2.3.1
bitarray==2.9.2
cattrs==23.2.3
certifi==2024.6.2
charset-normalizer==3.3.2
ckzg==1.0.2
cytoolz==0.12.3
dill==0.3.8
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.3.1
eth-utils==4.1.1
eth_abi==5.1.0
frozenlist==1.4.1
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isort==5.13.2
jedi==0.19.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
lsprotocol==2023.0.1
mccabe==0.7.0
multidict==6.0.5
packaging==24.1
parsimonious==0.10.0
parso==0.8.4
platformdirs==4.2.2
protobuf==5.27.2
psutil==6.0.0
pycodestyle==2.12.0
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pygls==1.3.1
pylint==3.2.5
python-bitcoinlib==0.12.2
python-dotenv==1.0.1
pyunormalize==15.1.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rlp==4.0.1
rpds-py==0.18.1
ruff==0.5.1
ruff-lsp==0.0.54
sniffio==1.3.1
tomlkit==0.12.5
toolz==0.12.1
typing_extensions==4.12.2
urllib3==2.2.2
web3==6.20.0
websockets==12.0
xxhash==3.4.1
yarl==1.9.4
